xmhf: libxmhfhw: vmx: asm(:::) --> asm volatile(); tested with syscalllog.

+++ xmhf-libs/libxmhfhw/include/xmhfhw.h
@@ -235,7 +235,7 @@ void xmhfhw_cpu_x86vmx_vmwrite(u32 encoding, u32 value);
 u32 xmhfhw_cpu_x86vmx_vmread(u32 encoding);
 u32 __vmx_vmclear(u64 vmcs);
 u32 __vmx_vmptrld(u64 vmcs);
-u32 __vmx_invvpid(int invalidation_type, u16 vpid, u32 linearaddress);
+u32 __vmx_invvpid(int invalidation_type, u32 vpid, u32 linearaddress);
 void __vmx_invept(u64 invalidation_type, u64 eptp);
 
 #endif //__ASSEMBLY__
